Raise ValueError if framebuffer file is empty
render.py now raises a ValueError if the framebuffer file is empty, indicating that the simulation has run but not enough clock cycles were elapsed for a whole frame to be written. The ValueError exception exits the Python script, propagates throughout PowerShell, and will be seen in ModelSim as an error.